add --options flag
The options flag take a file as an argument. You can specify --options
multiple times and the given files will be appended to a list of options
files.
The options file should contain flags separated by a newline.
You can specify --options in that file and it will recursively scan
--options given in other options files.
Functionally, it works like if --options would be replaced with the
content in the file.